requirements.txt
requests
schedule
resend
python-dotenv
